backend/requirements.txt
Django==5.0.2
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
django-environ==0.11.2
django-redis==5.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
environ==1.0
idna==3.6
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
overrides==7.4.0
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
drf-standardized-errors==0.12.6
